--- packages/EasyCodingStandard/src/Yaml/CheckerServiceParameterShifter.php
@@ -1,174 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\EasyCodingStandard\Yaml;
-
-use Nette\Utils\Strings;
-use ReflectionClass;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-
-/**
- * Before:
- *
- * services:
- *      # fixer
- *      ArrayFixer:
- *          syntax: short
- *      # sniff
- *      ArraySniff:
- *          syntax: short
- *
- * After:
- *
- * services:
- *      # fixer
- *      ArrayFixer:
- *          calls:
- *              - ['configure', [['syntax' => 'short']]
- *      # sniff
- *      ArraySniff:
- *          parameters:
- *              $syntax: 'short'
- */
-final class CheckerServiceParameterShifter
-{
-    /**
-     * @var CheckerConfigurationGuardian
-     */
-    private $checkerConfigurationGuardian;
-
-    /**
-     * @var string[]
-     */
-    private $serviceKeywords = [];
-
-    public function __construct()
-    {
-        $this->checkerConfigurationGuardian = new CheckerConfigurationGuardian();
-    }
-
-    /**
-     * @param mixed[] $yaml
-     * @return mixed[]
-     */
-    public function moveArgumentsToPropertiesOrMethodCalls(array $services): array
-    {
-        foreach ($services as $checker => $serviceDefinition) {
-            if (! $this->isCheckersClass($checker) || empty($serviceDefinition)) {
-                continue;
-            }
-
-            if (Strings::endsWith($checker, 'Fixer')) {
-                $services = $this->processFixer($services, $checker, $serviceDefinition);
-            }
-
-            if (Strings::endsWith($checker, 'Sniff')) {
-                $services = $this->processSniff($services, $checker, $serviceDefinition);
-            }
-
-            // cleanup parameters
-            foreach ($serviceDefinition as $key => $value) {
-                if ($this->isReservedKey($key)) {
-                    continue;
-                }
-
-                unset($services[$checker][$key]);
-            }
-        }
-
-        return $services;
-    }
-
-    private function isCheckersClass(string $checker): bool
-    {
-        return Strings::endsWith($checker, 'Fixer') || Strings::endsWith($checker, 'Sniff');
-    }
-
-    /**
-     * @param mixed[] $yaml
-     * @param mixed[] $serviceDefinition
-     * @return mixed[]
-     */
-    private function processFixer(array $services, string $checker, array $serviceDefinition): array
-    {
-        $this->checkerConfigurationGuardian->ensureFixerIsConfigurable($checker, $serviceDefinition);
-
-        foreach ($serviceDefinition as $key => $value) {
-            if ($this->isReservedKey($key)) {
-                continue;
-            }
-
-            $services[$checker]['calls'] = [
-                ['configure', [$serviceDefinition]],
-            ];
-        }
-
-        return $services;
-    }
-
-    /**
-     * @param mixed[] $yaml
-     * @param mixed[] $serviceDefinition
-     * @return mixed[]
-     */
-    private function processSniff(array $yaml, string $checker, array $serviceDefinition): array
-    {
-        // move parameters to property setters
-        foreach ($serviceDefinition as $key => $value) {
-            if ($this->isReservedKey($key)) {
-                continue;
-            }
-
-            $this->checkerConfigurationGuardian->ensurePropertyExists($checker, $key);
-            $services[$checker]['properties'][$key] = $this->escapeValue($value);
-        }
-
-        return $yaml;
-    }
-
-    /**
-     * @param mixed $value
-     * @return mixed
-     */
-    private function escapeValue($value)
-    {
-        if (is_bool($value) || is_numeric($value)) {
-            return $value;
-        }
-
-        if (is_array($value)) {
-            foreach ($value as $key => $nestedValue) {
-                $value[$key] = $this->escapeValue($nestedValue);
-            }
-
-            return $value;
-        }
-
-        return Strings::replace($value, '#@#', '@@');
-    }
-
-    /**
-     * @param string|int|bool $key
-     */
-    private function isReservedKey($key): bool
-    {
-        if (! is_string($key)) {
-            return false;
-        }
-
-        return in_array($key, $this->getServiceKeywords(), true);
-    }
-
-    /**
-     * @return string[]
-     */
-    private function getServiceKeywords(): array
-    {
-        if ($this->serviceKeywords) {
-            return $this->serviceKeywords;
-        }
-
-        $reflectionClass = new ReflectionClass(YamlFileLoader::class);
-
-        return $this->serviceKeywords = $reflectionClass->getStaticProperties()['serviceKeywords'];
-    }
-}
